+++ everest.yaml
@@ -1,5 +1,5 @@
 - Name: auspicioushelper
-  Version: 0.4.3.1
+  Version: 0.4.3.2
   DLL: bin/auspicioushelper.dll
   Dependencies:
     - Name: EverestCore
